Try adding package without barrel file
