+++ packages/data/README.md
@@ -1,0 +1,3 @@
+# Data
+
+Data
