+++ packages/domain-search/package.json
@@ -44,6 +44,7 @@
 		"storybook:start": "storybook dev"
 	},
 	"dependencies": {
+		"@automattic/data": "workspace:^",
 		"@tanstack/react-query": "^5.83.0",
 		"@wordpress/components": "^29.9.0",
 		"@wordpress/icons": "^10.23.0",
