+++ packages/domain-search/tsconfig-cjs.json
@@ -1,7 +1,7 @@
 {
 	"extends": "./tsconfig-build",
 	"compilerOptions": {
-		"module": "commonjs",
+		"module": "preserve",
 		"outDir": "dist/cjs"
 	}
 }
