+++ packages/domain-search/tsconfig.json
@@ -4,7 +4,8 @@
 		"declarationDir": "dist/types",
 		"outDir": "dist/esm",
 		"rootDir": "src",
-		"types": [ "jest", "@testing-library/jest-dom" ]
+		"types": [ "jest", "@testing-library/jest-dom" ],
+		"moduleResolution": "bundler"
 	},
 	"include": [ "src" ],
 	"references": [ { "path": "../components" } ]
